Move deployer to Utilities folder
Use StepGenerator in Deployer
